+++ engine/code/deepmind/dm_local.c
@@ -153,11 +153,11 @@ int dmlab_reward_override(const char* reason_opt, int player_id, int team,
       (intptr_t)score, 0, 0, 0, 0, 0, 0);
 }
 
-void dmlab_player_state(const playerState_t* ps, int team_score,
-                        int other_team_score) {
+void dmlab_player_state(const playerState_t* ps, const float eyePos[3],
+                        int team_score, int other_team_score) {
   trap_DeepmindCallback(DEEPMIND_SET_PLAYER_STATE, (intptr_t)ps,
-                        (intptr_t)team_score, (intptr_t)other_team_score, 0, 0,
-                        0, 0, 0, 0, 0, 0, 0);
+                        (intptr_t)eyePos, (intptr_t) team_score,
+                        (intptr_t)other_team_score, 0, 0, 0, 0, 0, 0, 0, 0);
 }
 
 int dmlab_make_screen_messages(int screen_width, int screen_height,